requirements.txt
mysql-connector-python
pyjwt
schedule
